Add method for sphere intersections
Added method for sphere intersection.
Source: Arvo, James, A Simple Method for Box-Sphere Intersection
Testing, p. 335-339, code: p. 730-732
For code sample, see:
http://tog.acm.org/resources/GraphicsGems/gems/BoxSphere.c
This is currently the best generally accepted method for this type of
intersection. There is a version of this algorithm out there with early
rejection, but whatever. This one is pretty well known.